+++ packages/relay-kit/src/packs/safe-4337/SafeOperationV06.ts
@@ -5,8 +5,8 @@ import {
   SafeUserOperation,
   SafeOperationOptions
 } from '@safe-global/types-kit'
-import SafeOperationBase from './SafeOperationBase'
-import { EIP712_SAFE_OPERATION_TYPE_V06 } from './constants'
+import SafeOperationBase from '@safe-global/relay-kit/packs/safe-4337/SafeOperationBase'
+import { EIP712_SAFE_OPERATION_TYPE_V06 } from '@safe-global/relay-kit/packs/safe-4337/constants'
 
 class SafeOperationV06 extends SafeOperationBase {
   userOperation!: UserOperationV06
