+++ packages/relay-kit/src/packs/safe-4337/SafeOperationV07.ts
@@ -5,8 +5,8 @@ import {
   SafeOperationOptions
 } from '@safe-global/types-kit'
 import { concat, Hex, isAddress, pad, toHex } from 'viem'
-import SafeOperationBase from './SafeOperationBase'
-import { EIP712_SAFE_OPERATION_TYPE_V07 } from './constants'
+import SafeOperationBase from '@safe-global/relay-kit/packs/safe-4337/SafeOperationBase'
+import { EIP712_SAFE_OPERATION_TYPE_V07 } from '@safe-global/relay-kit/packs/safe-4337/constants'
 import Safe from '@safe-global/protocol-kit'
 
 class SafeOperationV07 extends SafeOperationBase {
