+++ packages/relay-kit/src/packs/safe-4337/utils/signing.ts
@@ -1,104 +1,5 @@
-import {
-  Hex,
-  PublicRpcSchema,
-  createPublicClient,
-  encodeFunctionData,
-  encodePacked,
-  http,
-  rpcSchema,
-  toHex
-} from 'viem'
-import {
-  OperationType,
-  MetaTransactionData,
-  UserOperation,
-  UserOperationV07
-} from '@safe-global/types-kit'
-import {
-  EthSafeSignature,
-  encodeMultiSendData,
-  buildSignatureBytes
-} from '@safe-global/protocol-kit'
-import { ABI } from './constants'
-import {
-  BundlerClient,
-  RpcSchemaEntry,
-  Safe4337RpcSchema,
-  UserOperationStringValues
-} from './types'
-import { isEntryPointV7 } from './utils/entrypoint'
-
-/**
- * Gets the EIP-4337 bundler provider.
- *
- * @param {string} bundlerUrl The EIP-4337 bundler URL.
- * @return {BundlerClient} The EIP-4337 bundler provider.
- */
-export function createBundlerClient<ProviderCustomRpcSchema extends RpcSchemaEntry[] = []>(
-  bundlerUrl: string
-): BundlerClient<ProviderCustomRpcSchema> {
-  const provider = createPublicClient({
-    transport: http(bundlerUrl),
-    rpcSchema: rpcSchema<[...PublicRpcSchema, ...Safe4337RpcSchema, ...ProviderCustomRpcSchema]>()
-  })
-
-  return provider
-}
-
-/**
- * Encodes multi-send data from transactions batch.
- *
- * @param {MetaTransactionData[]} transactions - an array of transaction to to be encoded.
- * @return {string} The encoded data string.
- */
-export function encodeMultiSendCallData(transactions: MetaTransactionData[]): string {
-  return encodeFunctionData({
-    abi: ABI,
-    functionName: 'multiSend',
-    args: [
-      encodeMultiSendData(
-        transactions.map((tx) => ({ ...tx, operation: tx.operation ?? OperationType.Call }))
-      ) as Hex
-    ]
-  })
-}
-
-/**
- * Converts various bigint values from a UserOperation to their hexadecimal representation.
- *
- * @param {UserOperation} userOperation - The UserOperation object whose values are to be converted.
- * @returns {UserOperation} A new UserOperation object with the values converted to hexadecimal.
- */
-export function userOperationToHexValues(
-  userOperation: UserOperation,
-  entryPointAddress: string
-): UserOperationStringValues {
-  const userOpV07 = userOperation as UserOperationV07
-
-  const userOperationWithHexValues = {
-    ...userOperation,
-    nonce: toHex(BigInt(userOperation.nonce)),
-    callGasLimit: toHex(userOperation.callGasLimit),
-    verificationGasLimit: toHex(userOperation.verificationGasLimit),
-    preVerificationGas: toHex(userOperation.preVerificationGas),
-    maxFeePerGas: toHex(userOperation.maxFeePerGas),
-    maxPriorityFeePerGas: toHex(userOperation.maxPriorityFeePerGas),
-    ...(isEntryPointV7(entryPointAddress)
-      ? {
-          paymaster: userOpV07.paymaster !== '0x' ? userOpV07.paymaster : null,
-          paymasterData: userOpV07.paymasterData !== '0x' ? userOpV07.paymasterData : null,
-          paymasterVerificationGasLimit: userOpV07.paymasterVerificationGasLimit
-            ? toHex(userOpV07.paymasterVerificationGasLimit)
-            : null,
-          paymasterPostOpGasLimit: userOpV07.paymasterPostOpGasLimit
-            ? toHex(userOpV07.paymasterPostOpGasLimit)
-            : null
-        }
-      : {})
-  }
-
-  return userOperationWithHexValues
-}
+import { Hex, encodePacked, toHex } from 'viem'
+import { EthSafeSignature, buildSignatureBytes } from '@safe-global/protocol-kit'
 
 /**
  * Passkey Dummy client data JSON fields. This can be used for gas estimations, as it pads the fields enough
